revising how we handle handlers (#1236)
