+++ app/contacts/contact_websocket.py
@@ -2,36 +2,33 @@ import asyncio
 
 import websockets
 
-from app.contacts.handlers.h_websocket import Handler
 from app.utility.base_world import BaseWorld
 
 
 class WebSocket(BaseWorld):
 
     def __init__(self, services):
-        self.name = 'web socket'
+        self.name = 'websocket'
         self.description = 'Accept data through web sockets'
         self.log = self.create_logger('contact_websocket')
-        self.socket_handler = SocketHandler(services)
+        self.handler = Handler(services)
 
     async def start(self):
         loop = asyncio.get_event_loop()
         web_socket = self.get_config('app.contact.websocket')
-        loop.create_task(await websockets.serve(self.socket_handler.handle, '0.0.0.0', web_socket.split(':')[1]))
-
-    @staticmethod
-    def valid_config():
-        return True
+        loop.create_task(await websockets.serve(self.handler.handle, '0.0.0.0', web_socket.split(':')[1]))
 
 
-class SocketHandler:
+class Handler:
 
     def __init__(self, services):
         self.services = services
-        self.log = BaseWorld.create_logger('websocket_session')
+        self.handles = []
+        self.log = BaseWorld.create_logger('websocket_handler')
 
     async def handle(self, socket, path):
         try:
-            await Handler(path.split('/')[1]).handle(socket, path, self.services)
+            for handle in [h for h in self.handles if h.tag == path.split('/')[1]]:
+                await handle.run(socket, path, self.services)
         except Exception as e:
             self.log.debug(e)
